flask_api/requirements.txt
a2wsgi==1.10.7
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
connexion==3.1.0
Django==5.1.4
djangorestframework==3.15.2
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mysqlclient==2.2.6
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
swagger_ui_bundle==1.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.1.3
